Rewrite tests with Google Test
